Arduino Code
This code has an LCD added that displays each piece of data at a time,
as well as a countdown to when the data will be sent through the ESP8266
module.

I haven't yet tested it with the ESP8266 module but the LCD works fine.
No modifications to the code for sending the data so I'd imagine it
shouldn't be too hard to debug if anything goes wrong.